# docs/swagger.yaml
basePath: /
definitions:
...
  N1ktarchik_Board_of_issues_internal_core_domain.User:
    properties:
      created_at:
        example: "2026-01-01T00:00:00Z"
        type: string
      email:
        example: bonya123@example.com
        type: string
      id:
        example: 636e856-e12b-56d9-f987-333222561234
        type: string
      login:
        example: bonya123
        type: string
      name:
        example: Bonya
        type: string
      password:
        example: pass2000
        type: string
    type: object
  N1ktarchik_Board_of_issues_internal_core_transport_response.ErrorResponse:
    properties:
      code:
        example: 0
        type: integer
      error:
        example: any string-code of error
        type: string
      message:
        example: any error message
        type: string
    type: object
...
  internal_features_users_transport_http.UsersRequestDTO:
    properties:
      email:
        example: bonya123@example.com
        type: string
      login:
        example: bonya123
        type: string
      name:
        example: Bonya
        type: string
      password:
        example: pass2000
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: klejmenov663@email.com
    name: Nikita Kleymenov
    url: https://t.me/n1ktarchik
  description: REST API for the Board of Issues task management system.
  title: Board Of Issues API
  version: "2.0"
paths:
...
  /users/update:
    patch:
      consumes:
      - application/json
      description: Update current user's name or email or password
      parameters:
      - description: New User Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_features_users_transport_http.UsersRequestDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated user information
          schema:
            $ref: '#/definitions/N1ktarchik_Board_of_issues_internal_core_domain.User'
        "400":
          description: 'Possible: invalid_email, bad_request'
          schema:
            $ref: '#/definitions/N1ktarchik_Board_of_issues_internal_core_transport_response.ErrorResponse'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/N1ktarchik_Board_of_issues_internal_core_transport_response.ErrorResponse'
        "404":
          description: user_not_found
          schema:
            $ref: '#/definitions/N1ktarchik_Board_of_issues_internal_core_transport_response.ErrorResponse'
        "409":
          description: email_already_exists
          schema:
            $ref: '#/definitions/N1ktarchik_Board_of_issues_internal_core_transport_response.ErrorResponse'
        "500":
          description: internal_server_error
          schema:
            $ref: '#/definitions/N1ktarchik_Board_of_issues_internal_core_transport_response.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update user profile
      tags:
      - users
securityDefinitions:
  ApiKeyAuth:
    description: 'Enter the token in the format: Bearer <JWT_TOKEN>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"